src/requirements.txt
python-i18n
pymysql
PyYAML